agentic_backtesting/agentcore-enablement/requirements_fixed.txt
bedrock-agentcore>=1.0.0
strands-agents>=1.0.0
pandas>=1.5.0,<3.0.0
numpy>=1.24.0,<2.0.0
backtrader==1.9.78.123
boto3>=1.26.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
httpx>=0.24.0,<1.0.0
bedrock-agentcore-starter-toolkit>=0.1.0,<1.0.0
pyiceberg[s3fs]>=0.5.0,<1.0.0
